Move interfaces and types inside mcrypt.d.ts
