--- types/js-rijndael/Cipher.d.ts
@@ -1,8 +1,0 @@
-export type Cipher = Record<CipherAlgorithm, CipherBody>;
-
-export type CipherAlgorithm = 'rijndael-128' | 'rijndael-192' | 'rijndael-256';
-
-export interface CipherBody {
-    blockSize: number;
-    keySize: number;
-}
