--- types/js-rijndael/Mode.d.ts
@@ -1,1 +1,0 @@
-export type Mode = 'cbc' | 'cfb' | 'ctr' | 'ecb' | 'ncfb' | 'nofb';
