+++ types/js-rijndael/mcrypt.d.ts
@@ -1,5 +1,13 @@
-import { CipherBody, CipherAlgorithm } from './Cipher';
-import { Mode } from './Mode';
+export type Cipher = Record<CipherAlgorithm, CipherBody>;
+
+export type CipherAlgorithm = 'rijndael-128' | 'rijndael-192' | 'rijndael-256';
+
+export interface CipherBody {
+    blockSize: number;
+    keySize: number;
+}
+
+export type Mode = 'cbc' | 'cfb' | 'ctr' | 'ecb' | 'ncfb' | 'nofb';
 
 declare const ciphers: CipherBody & {
     'rijndael-128': { blockSize: 16; keySize: 32 };
